pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.14",
    "apscheduler>=3.11.0",
    "cryptography>=45.0.5",
    "flask>=3.1.1",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
]
